Move samplers to avoid circular imports. Rename pretrain back
