+++ examples/sinc/sinc_nas.py
@@ -12,7 +12,7 @@ from syne_tune.config_space import randint
 from syne_tune.report import Reporter
 from torch.utils.data import DataLoader
 
-from whittle.sampling.samplers import RandomSampler
+from whittle.sampling.random_sampler import RandomSampler
 from whittle.training_strategies import SandwichStrategy
 
 from .model import MLP
