+++ whittle/sampling/__init__.py
@@ -1,5 +1,5 @@
 from __future__ import annotations
 
-from .samplers import Samplers
+from .samplers import Samplers, get_sampler
 
-__all__ = ["Samplers"]
+__all__ = ["Samplers", "get_sampler"]
